ETCM-161: Move the kademlia package into its own submodule.
