+++ scalanet/kademlia/test/src/io/iohk/scalanet/kademlia/KRouterSpec.scala
@@ -1,19 +1,19 @@
-package io.iohk.scalanet.peergroup.kademlia
+package io.iohk.scalanet.kademlia
 
 import java.time.Clock
 import java.util.UUID
 
 import cats.effect.concurrent.Ref
-import io.iohk.scalanet.peergroup.kademlia.Generators.{aRandomBitVector, aRandomNodeRecord}
-import io.iohk.scalanet.peergroup.kademlia.KMessage.KRequest.{FindNodes, Ping}
-import io.iohk.scalanet.peergroup.kademlia.KMessage.{KRequest, KResponse}
-import io.iohk.scalanet.peergroup.kademlia.KMessage.KResponse.{Nodes, Pong}
-import io.iohk.scalanet.peergroup.kademlia.KRouter.{Config, NodeRecord}
-import io.iohk.scalanet.peergroup.kademlia.KRouterSpec.KNetworkScalanetInternalTestImpl.{
+import io.iohk.scalanet.kademlia.Generators.{aRandomBitVector, aRandomNodeRecord}
+import io.iohk.scalanet.kademlia.KMessage.KRequest.{FindNodes, Ping}
+import io.iohk.scalanet.kademlia.KMessage.{KRequest, KResponse}
+import io.iohk.scalanet.kademlia.KMessage.KResponse.{Nodes, Pong}
+import io.iohk.scalanet.kademlia.KRouter.{Config, NodeRecord}
+import io.iohk.scalanet.kademlia.KRouterSpec.KNetworkScalanetInternalTestImpl.{
   KNetworkScalanetInternalTestImpl,
   NodeData
 }
-import io.iohk.scalanet.peergroup.kademlia.KRouterSpec._
+import io.iohk.scalanet.kademlia.KRouterSpec._
 import monix.eval.Task
 import monix.execution.Scheduler
 import monix.reactive.Observable
